+++ build_runner_core/lib/build_runner_core.dart
@@ -4,7 +4,6 @@
 
 export 'package:build/build.dart' show PostProcessBuildStep, PostProcessBuilder;
 
-export 'src/asset/batch.dart' show wrapInBatch;
 export 'src/asset/file_based.dart';
 export 'src/asset/finalized_reader.dart';
 export 'src/asset/writer.dart';